refactor: package by domain after packaging by layer
